refactor(api): extract core from client.ts
